+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for fakedata
 
+## 0.2.2
+
+* Add meepoQuote function to the module Faker.Game.Dota
+* Add Faker.Music.Opera module
+
 ## 0.2.1
 
 * Fix bug in Book module